fix(routes): Se agrego seguridad para que no se pueda hacer peticiones de edición y eliminación de fuentes externas.

+++ front/src/assets/services/deleteProduct.js
@@ -1,13 +1,24 @@
 import { BASEURL } from "../db/connectURL";
+import { obtenerDelSessionStorage } from "../utils/localStorage";
 
 export async function handleDelete(producto, onProductUpdate) {
   const productoID = producto._id;
+
+  const token = obtenerDelSessionStorage("token");
+
+  if (!token) {
+    console.error(
+      "Token no encontrado. Debes iniciar sesion como administrador para poder hacer peticiones"
+    );
+    return false;
+  }
 
   try {
     const response = await fetch(`${BASEURL}/api/productos/${productoID}`, {
       method: "DELETE",
       headers: {
         "Content-Type": "application/json",
+        Authorization: `Bearer ${token}`,
       },
     });
